requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
gspread==6.1.4
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.28.0
numpy==2.2.3
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
six==1.17.0
smmap==5.0.2
streamlit==1.42.2
streamlit_carousel==1.1.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
twilio==9.4.6
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
yarl==1.18.3
